refactor(#159): shred/dto 생성하여 유효성 검사로직 추가
