+++ ai/src/main/java/dukku/ai/app/usecase/CreateAiMemoryUseCase.java
@@ -3,8 +3,8 @@ package dukku.ai.app.usecase;
 import org.springframework.stereotype.Service;
 
 import dukku.ai.entity.AiMemory;
-import dukku.ai.entity.enums.MemorySubType;
-import dukku.ai.entity.enums.MemoryType;
+import dukku.common.shared.ai.type.MemorySubType;
+import dukku.common.shared.ai.type.MemoryType;
 import dukku.ai.out.AiMemoryRepository;
 
 @Service
