+++ common/src/main/java/dukku/common/shared/ai/dto/ChatResponse.java
@@ -1,0 +1,7 @@
+package dukku.common.shared.ai.dto;
+
+public record ChatResponse(
+        String conversationId,
+        String reply
+) {
+}
