+++ common/src/main/java/dukku/common/shared/ai/dto/UpdateAiMemoryRequest.java
@@ -1,0 +1,7 @@
+package dukku.common.shared.ai.dto;
+
+public record UpdateAiMemoryRequest(
+        Double importanceScore,
+        Double confidenceScore
+) {
+}
